+++ server/controller/BookingController.js
@@ -1,5 +1,6 @@
 import { sendErrorResponse, sendOkResponse } from "../core/responses.js";
 import BookingSchema from "../models/AccessoriesSchema.model.js";
+import mongoose from "mongoose";
 
 export const getBookingsVehicle = async (req, res) => {
     try {
@@ -16,6 +17,9 @@ export const getBookingsVehicle = async (req, res) => {
 
 export const getBookingVehicle = async (req, res) => {
     try {
+        if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+            return sendErrorResponse(res, { message: `Invalid ID format` }, 400)
+        }
         const Booking = await BookingSchema.findById(req.params.id);
         if (!Booking) {
             sendErrorResponse(res, { message: `User ID: ${req.params.id} not found` }, 404)
@@ -55,6 +59,9 @@ export const updateBookingVehicle = async (req, res) => {
 
 export const deleteBookingVehicle = async (req, res) => {
     try {
+        if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+            return sendErrorResponse(res, { message: `Invalid ID format` }, 400)
+        }
         const Booking = await BookingSchema.findByIdAndDelete(req.params.id);
         if (!Booking) {
             sendErrorResponse(res, { message: `Data not found` }, 404)
