+++ server/controller/TenantController.js
@@ -1,5 +1,6 @@
 import { sendErrorResponse, sendOkResponse } from "../core/responses.js";
 import TenantSchema from "../models/TenantSchema.model.js";
+import mongoose from "mongoose";
 
 export const getTenantsVehicle = async (req, res) => {
     try {
@@ -16,6 +17,9 @@ export const getTenantsVehicle = async (req, res) => {
 
 export const getTenantVehicle = async (req, res) => {
     try {
+        if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+            return sendErrorResponse(res, { message: `Invalid ID format` }, 400)
+        }
         const Tenant = await TenantSchema.findById(req.params.id);
         if (!Tenant) {
             sendErrorResponse(res, { message: `User ID: ${req.params.id} not found` }, 404)
@@ -55,6 +59,9 @@ export const updateTenantVehicle = async (req, res) => {
 
 export const deleteTenantVehicle = async (req, res) => {
     try {
+        if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+            return sendErrorResponse(res, { message: `Invalid ID format` }, 400)
+        }
         const Tenant = await TenantSchema.findByIdAndDelete(req.params.id);
         if (!Tenant) {
             sendErrorResponse(res, { message: `Data not found` }, 404)
